Adds refactoring.
