+++ project/src/components/review-form/review-form.jsx
@@ -3,19 +3,24 @@ import PropTypes from 'prop-types';
 import {connect} from 'react-redux';
 import {postReview} from '../../store/api-actions';
 import RatingList from '../rating-list/rating-list';
+import ReviewComment from '../review-comment/review-comment';
 
 const MIN_CHARS_COUNT = 50;
+const MAX_CHARS_COUNT = 300;
 
 function ReviewForm({roomId, sendReview}) {
   // const [review, setReview] = useState({rating: null, comment: ''});
   const [rating, setRating] = useState(null);
-  const [review, setReview] = useState('');
+  const [comment, setComment] = useState('');
+
+  const isButtonDisabled = rating === null || comment.length < MIN_CHARS_COUNT
+  || comment.length > MAX_CHARS_COUNT;
 
   const onFormSubmit = (evt) => {
     evt.preventDefault();
-    sendReview({id: roomId, comment: review, rating: rating});
+    sendReview({id: roomId, comment: comment, rating: rating});
     setRating(null);
-    setReview('');
+    setComment('');
   };
 
   return (
@@ -24,20 +29,13 @@ function ReviewForm({roomId, sendReview}) {
       method="post"
       onSubmit={(evt) => onFormSubmit(evt)}
     >
-      <label className="reviews__label form__label"
+      <label
+        className="reviews__label form__label"
         htmlFor="review"
       >Your review
       </label>
       <RatingList rating={rating} setRating={setRating} />
-      <textarea
-        className="reviews__textarea form__textarea"
-        id="review"
-        name="review"
-        placeholder="Tell how was your stay, what you like and what can be improved"
-        value={review}
-        onChange={(evt) => setReview(evt.target.value)}
-      >
-      </textarea>
+      <ReviewComment comment={comment} setComment={setComment} />
       <div className="reviews__button-wrapper">
         <p className="reviews__help">
           To submit review please make sure to set <span className="reviews__star">rating</span> and describe your stay with at least <b className="reviews__text-amount">50 characters</b>.
@@ -45,7 +43,7 @@ function ReviewForm({roomId, sendReview}) {
         <button
           className="reviews__submit form__submit button"
           type="submit"
-          disabled={review.length < MIN_CHARS_COUNT && true}
+          disabled={isButtonDisabled ? 'disabled' : ''}
         >Submit
         </button>
       </div>
@@ -54,7 +52,9 @@ function ReviewForm({roomId, sendReview}) {
 }
 
 ReviewForm.propTypes = {
-  roomId: PropTypes.number.isRequired,
+  roomId: PropTypes.oneOfType([
+    PropTypes.string,
+    PropTypes.number]),
   sendReview: PropTypes.func.isRequired,
 };
